requirements.txt
dash==0.26.3  # The core dash backend
dash-renderer==0.13.2  # The dash front-end
dash-html-components==0.11.0  # HTML components
dash-core-components==0.28.1  # Supercharged components
plotly # Plotly graphing library used in examples
backcall==0.1.0
cached-property==1.4.3
chardet==3.0.4
decorator==4.3.0
EasyProcess==0.2.3
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.7
itsdangerous==0.24
jsonschema==2.6.0
MarkupSafe==1.0
nbformat==4.4.0
numpy==1.15.1
pandas==0.23.4
parso==0.3.1
patool==1.12
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
retry==0.9.2
retrying==1.3.3
simplegeneric==0.8.1
six==1.11.0
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
